+++ weui/CHANGELOG.md
@@ -1,3 +1,17 @@
-## [0.0.1] - TODO: Add release date.
+## 0.0.4
 
-* TODO: Describe initial release.
+### 修改
+- WeInput添加height参数, 设置高度
+- WeInput添加maxLength参数, 限制最大输入数量
+- WeInput添加focusNode参数, 原生FocusNode
+- WeForm添加spacing参数, 设置表单内做边空白间距
+- WeForm添加boxBorder参数, 控制表单外层是否显示边框
+- WeToast.fail方法添加onClose回掉
+- WeToast.success方法添加onClose回掉
+### 新增
+- 新增WeTheme类, 全局控制主题
+- 新增WeCheckbox组件
+- 新增WeRadio组件
+### 修复
+- 修复WeInput组件多行输入时无法点击键盘的换行
+- 修复WeGrid组件中itemBuilder没有返回正确的index